feat: restructure server codebase by implementing new data access layer and API structure; add Prettier configuration for code formatting
