--- server/datastore/LikesDao.ts
@@ -1,4 +1,0 @@
-import { Likes } from "../src/types.js";
-export interface LikesDao {
-createLike(like : Likes) : void;
-}
