--- server/datastore/UserDao.ts
@@ -1,6 +1,0 @@
-import { User } from "../src/types.js";
-export interface UesrDao{
-    createUser(user: User):void;
-    getUserByEmail(email: string): User | undefined;
-    getUserByUsername(Username: string): User | undefined; 
-}
