+++ server/nodemon.json
@@ -1,6 +1,6 @@
 {
-  "watch": ["src"],
-  "ext": "ts,json",
-  "ignore": ["dist"],
-  "exec": "node --loader ts-node/esm ./src/index.ts"
+    "watch": ["src"],
+    "ext": "ts,json",
+    "ignore": ["dist"],
+    "exec": "node --loader ts-node/esm ./src/index.ts"
 }
