+++ server/src/Handlers/index.ts
@@ -1,0 +1,1 @@
+export * from './PostHandler.js';
