+++ server/src/datastore/LikesDao.ts
@@ -1,0 +1,4 @@
+import { Likes } from '../types.js';
+export interface LikesDao {
+    createLike(like: Likes): void;
+}
