+++ server/src/datastore/PostDao.ts
@@ -1,0 +1,7 @@
+import { Post } from '../types.js';
+export interface PostDao {
+    listPosts(): Post[];
+    createPost(post: Post): void;
+    getPost(id: string): Post | undefined;
+    deletePost(id: string): void;
+}
