+++ server/src/datastore/UserDao.ts
@@ -1,0 +1,6 @@
+import { User } from '../types.js';
+export interface UesrDao {
+    createUser(user: User): void;
+    getUserByEmail(email: string): User | undefined;
+    getUserByUsername(Username: string): User | undefined;
+}
